test(dropdown): restore to original structure for screenshot tests
