+++ packages/core/src/components/dropdown/test/default/dropdown.e2e.ts-snapshots/tds-dropdown-default-clicking-the-dropdown-button-opens-the-dropdown-list-1-linux.png
