+++ packages/core/src/components/dropdown/test/default/dropdown.e2e.ts-snapshots/tds-dropdown-default-clicking-the-dropdown-opens-the-dropdown-list-then-click-Option-1-1-linux.png
