+++ packages/core/src/components/dropdown/test/default/dropdown.e2e.ts-snapshots/tds-dropdown-default-focusElement-method-focus-and-opens-the-dropdown-list-1-linux.png
