+++ packages/core/src/components/dropdown/test/default/dropdown.e2e.ts-snapshots/tds-dropdown-default-renders-default-dropdown-correctly-1-linux.png
